requirements.txt
playwright
openai
httpx
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pandas
python-dotenv
